dSYMs archiving as a build artifacts was fixed and refactored
By fact, previous zipping implementation was wrong and it was not possible to unarchive the build. Yikes..
It was fixed using http://www.mkyong.com/java/how-to-compress-files-in-zip-format/ as a reference.
Zipping routine was also moved to a Zip class in util subpackage.
It's still needed to test what would happen if we have more then one dSYM in a build.