Action was added
It would be nice to show itms-services:// links on the build results page. That's why action was added.
It's empty for now, but it should be possible to show itms-services links there.
It would also be needed to get some information from the built application's Info.plist file (like application's bundle ID, for example)