Build artifacts renaming was added
It's needed to provide a way to customize names which are used for build artifacts (.dSYM and .ipa).
That customizing might be handy when you, for instance, have more then one build step for the same sources.
It's done by introducing an artifacts template setting in a build configuration. This template is used when build artifacts are collected.
There's only one 'variable' available for now – $APP_NAME which might be used to put application's name, like, for example:
	- for MyApp.app.dSYM it would produce MyApp-production.app.dSYM with template "$APP_NAME-production"
	- for MyApp.ipa is would produce MyApp-appStore.ipa with template "$APP_NAME-appStore"
"$APP_NAME" is used by default as template. If empty string is provided as archive's template – it's treated as $APP_NAME as well.