Keychain path is used in xcodebuild, .keychain extension was added to a keychain name, /tmp is used as a temp directory
Seems like signing was failing on some machines. Probably because of wrong keychain specification.
Seems like extension adding is not too much. The full keychain file path is used in xcodebuild.
Temporary directory path was looking weird so, as we might assume that the build would be run on OS X, /tmp is used as a temp directory.