projectRootPath argument was added to installPods function and removed from iOSBuilderExecutor constructor, some refactoring was done
The only step which requires projectRoot to be specified is pods installing step. That's why we don't need that argument to be in the constructor.
Build-related exception stacktraces are now printed to the listener's logger – to simplify problems diagnosis.
EnvVars expanding was removed on some places as it's not mandatory to be there. For example, it's possible to remove them in build step command composing because all the envVars would be expanded during command executing.