Artifacts archiving was removed from the builder.
It's not builder's business to archive build artifacts as that could be done later in a post-build action.
Meanwhile it might be useful to specify a different path for an IPA archive and to zip dSYM folder. That's why two (in fact, three) new options were added.
In addition, zipping is possible using Jenkins' FilePath.zip method. That's why Zip class was removed from the plugin.