Code resigning was removed from the packaging step. Default IPA and dSYM zip names were changed.
It's almost useless to resign the IPA and, additionally, it's not possible to specify keychain during application packaging.
That's why it's more safe to remove resigning-related parameters from the packaging step.
One small change – the default name for ipa and dSYM's zip was changed from $APP_NAME to build/$APP_NAME – to place archives in the build folder.